requirements.txt
fastapi
python-multipart
jinja2
anthropic>=0.18.0
python-dotenv
uvicorn 